fix : 중복되는 타입 제거 (#59)
