+++ src/queries/houseMemo/useHouseMemoInfoById.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getHouseMemoById } from '@/services/house.memo.byId';
-import { ApiResponse } from '@/types/house-memo-byId';
+import { ApiResponse } from '@/types/house-memo';
 
 export function useHouseMemoInfoById(id: string) {
   return useQuery<ApiResponse>({
