+++ src/services/house.memo.byId.ts
@@ -1,7 +1,7 @@
-import { ApiResponse, HouseMemoData } from '@/types/house-memo-byId';
+import { ApiResponse, HouseMemoDataById } from '@/types/house-memo';
 import { api } from '@/lib/api/ky';
 
-export async function getHouseMemoById(id: string): Promise<ApiResponse<HouseMemoData>> {
-  const res = await api.get(`api/property/${id}`).json<ApiResponse<HouseMemoData>>();
+export async function getHouseMemoById(id: string): Promise<ApiResponse<HouseMemoDataById>> {
+  const res = await api.get(`api/property/${id}`).json<ApiResponse<HouseMemoDataById>>();
   return res;
 }
