+++ src/types/house-memo.ts
@@ -19,9 +19,39 @@ export type HouseType =
   | 'BOARDING'
   | 'ETC';
 
+export type ChecklistItem = {
+  itemId: number;
+  question: string;
+  description?: string;
+};
+
+export type ChecklistSection = {
+  categoryId: number;
+  categoryName: string;
+  memo?: string;
+  items: ChecklistItem[];
+};
+
+export type ChecklistCategory = {
+  order: number;
+  name: string;
+};
+
+export type Checklist = {
+  categories: ChecklistCategory[];
+  sections: ChecklistSection[];
+};
+
 export type ChecklistMemo = {
   categoryId: number;
   memo: string;
+};
+
+// types/image.ts
+export type Image = {
+  imageId: number;
+  url: string;
+  order: number;
 };
 
 export type HouseMemo = {
@@ -49,4 +79,58 @@ export type HouseMemoInputField = {
   placeholder?: string;
   required?: boolean;
   unit?: string;
+};
+
+export type HouseMemoByIdData = {
+  images: Image[];
+  propertyId: number;
+  planId: number;
+  planName: string;
+  folderId: number;
+  folderName: string;
+  feeling: Feeling;
+  propertyName: string;
+  memo?: string;
+  referenceUrl?: string | null;
+  address: string;
+  detailAddress?: string;
+  latitude: number;
+  longitude: number;
+  contractType: ContractType;
+  houseType: HouseType;
+  depositBig?: number;
+  depositSmall?: number;
+  managementFee?: number | null;
+  moveInInfo?: string;
+  checklist: Checklist;
+};
+
+export type HouseMemoDataById = {
+  images: Image[];
+  propertyId: number;
+  planId: number;
+  planName: string;
+  folderId: number;
+  folderName: string;
+  feeling: Feeling;
+  propertyName: string;
+  memo?: string;
+  referenceUrl?: string | null;
+  address: string;
+  detailAddress?: string;
+  latitude: number;
+  longitude: number;
+  contractType: ContractType;
+  houseType: HouseType;
+  depositBig?: number;
+  depositSmall?: number;
+  managementFee?: number | null;
+  moveInInfo?: string;
+  checklist: Checklist;
+};
+
+export type ApiResponse<T = HouseMemoDataById> = {
+  code: string;
+  message: string;
+  data: T | null;
 };
